backend/requirements.txt
fastapi
uvicorn
gunicorn
pandas
scikit-learn==1.8.0
joblib
mlflow
pydantic
python-dotenv
python-box